# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for Phase 1 user top-up campaigns (admin + user-facing). Operates under `/campaign-center-api/v1`;",
        "title": "Campaign Center API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/campaign-center-api/v1",
    "paths": {
...
        "/admin/landing-pages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-landing-page"
                ],
                "summary": "List landing pages (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Default language filter e.g. en",
                        "name": "defaultLang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-landing-page"
                ],
                "summary": "Create landing page (admin)",
                "parameters": [
                    {
                        "description": "Landing page content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LandingPageBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "validation error",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/landing-pages/{landingPageId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-landing-page"
                ],
                "summary": "Get landing page (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Landing page ID",
                        "name": "landingPageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Requested language (falls back to default)",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-landing-page"
                ],
                "summary": "Update landing page (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Landing page ID",
                        "name": "landingPageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Landing page content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LandingPageBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "409": {
                        "description": "not draft",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/landing-pages/{landingPageId}/detail/{lang}": {
            "get": {
                "description": "title/description/terms come from campaign_landing_page_translations when a row exists for lang; otherwise from campaign_landing_pages. bannerImageUrl, status, timestamps always from campaign_landing_pages.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-landing-page"
                ],
                "summary": "Get landing page detail by locale (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Landing page ID",
                        "name": "landingPageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Locale tag, e.g. ja, zh-CN",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/api.LandingPageLocaleDetailHTTPResponse"
                        }
                    },
                    "400": {
                        "description": "invalid path",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/landing-pages/{landingPageId}/publish": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-landing-page"
                ],
                "summary": "Publish landing page (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Landing page ID",
                        "name": "landingPageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Operator",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PublishOperatorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.LandingPageBody": {
            "type": "object",
            "required": [
                "bannerImageUrl",
                "defaultLang",
                "description",
                "terms",
                "title"
            ],
            "properties": {
                "bannerImageUrl": {
                    "type": "string"
                },
                "defaultLang": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "terms": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.LandingPageLocaleDetailData": {
            "type": "object",
            "properties": {
                "bannerImageUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "defaultLang": {
                    "type": "string",
                    "example": "en"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 2001
                },
                "lang": {
                    "type": "string",
                    "example": "ja"
                },
                "status": {
                    "type": "integer"
                },
                "terms": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "api.LandingPageLocaleDetailHTTPResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/api.LandingPageLocaleDetailData"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "api.PublishOperatorReq": {
            "type": "object",
            "required": [
                "operator"
            ],
            "properties": {
                "operator": {
                    "type": "string"
                }
            }
        },
...
        "data.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}